refactor(project): rename 'compat' pkg to 'test'
fixes #231
