+++ test/doc.go
@@ -1,4 +1,4 @@
-package compat
+package test
 
 // this package contains adoc file with various features that are verified to be supported by Libasciidoc
 // i.e., they should produce the same output as Asciidoctor
