+++ test/fixtures/pending/encoding.adoc
